+++ fractal_server/app/models/v2/__init__.py
@@ -6,8 +6,9 @@ from .collection_state import CollectionStateV2
 from .dataset import DatasetV2
 from .job import JobV2
 from .project import ProjectV2
-from .task import TaskGroupV2
 from .task import TaskV2
+from .task_group import TaskGroupActivityV2
+from .task_group import TaskGroupV2
 from .workflow import WorkflowV2
 from .workflowtask import WorkflowTaskV2
 
@@ -18,6 +19,7 @@ __all__ = [
     "ProjectV2",
     "CollectionStateV2",
     "TaskGroupV2",
+    "TaskGroupActivityV2",
     "TaskV2",
     "WorkflowTaskV2",
     "WorkflowV2",