+++ fractal_server/app/routes/api/v2/workflow.py
@@ -23,7 +23,7 @@ from ._aux_functions import _get_submitted_jobs_statement
 from ._aux_functions import _get_workflow_check_owner
 from ._aux_functions_tasks import _add_warnings_to_workflow_tasks
 from fractal_server.app.models import UserOAuth
-from fractal_server.app.models.v2.task import TaskGroupV2
+from fractal_server.app.models.v2 import TaskGroupV2
 from fractal_server.app.routes.auth import current_active_user
 
 router = APIRouter()