+++ fractal_server/app/schemas/v2/__init__.py
@@ -27,6 +27,9 @@ from .task_collection import CollectionStateReadV2  # noqa F401
 from .task_collection import CollectionStatusV2  # noqa F401
 from .task_collection import TaskCollectCustomV2  # noqa F401
 from .task_collection import TaskCollectPipV2  # noqa F401
+from .task_group import TaskGroupActivityActionV2  # noqa F401
+from .task_group import TaskGroupActivityStatusV2  # noqa F401
+from .task_group import TaskGroupActivityV2Read  # noqa F401
 from .task_group import TaskGroupCreateV2  # noqa F401
 from .task_group import TaskGroupReadV2  # noqa F401
 from .task_group import TaskGroupUpdateV2  # noqa F401