+++ fractal_server/tasks/v2/collection_ssh.py
@@ -2,16 +2,15 @@ import os
 from pathlib import Path
 from tempfile import TemporaryDirectory
 
-from sqlalchemy.orm.attributes import flag_modified
-
 from .database_operations import create_db_tasks_and_update_task_group
 from .utils_background import _handle_failure
 from .utils_background import _prepare_tasks_metadata
 from .utils_background import _set_collection_state_data_status
+from .utils_background import _set_task_group_activity_status
 from fractal_server.app.db import get_sync_db
-from fractal_server.app.models.v2 import CollectionStateV2
 from fractal_server.app.models.v2 import TaskGroupV2
 from fractal_server.app.schemas.v2 import CollectionStatusV2
+from fractal_server.app.schemas.v2 import TaskGroupActivityStatusV2
 from fractal_server.app.schemas.v2.manifest import ManifestV2
 from fractal_server.config import get_settings
 from fractal_server.logger import get_logger
@@ -82,6 +81,7 @@ def _customize_and_run_template(
 def collect_package_ssh(
     *,
     state_id: int,
+    task_group_activity_id: int,
     task_group: TaskGroupV2,
     fractal_ssh: FractalSSH,
     tasks_base_dir: str,
@@ -165,8 +165,15 @@ def collect_package_ssh(
                     logger_name=LOGGER_NAME,
                     db=db,
                 )
+                _set_task_group_activity_status(
+                    task_group_activity_id=task_group_activity_id,
+                    new_status=TaskGroupActivityStatusV2.ONGOING,
+                    logger_name=LOGGER_NAME,
+                    db=db,
+                )
                 _refresh_logs(
                     state_id=state_id,
+                    task_group_activity_id=task_group_activity_id,
                     log_file_path=log_file_path,
                     db=db,
                 )
@@ -183,6 +190,7 @@ def collect_package_ssh(
                 remove_venv_folder_upon_failure = True
                 _refresh_logs(
                     state_id=state_id,
+                    task_group_activity_id=task_group_activity_id,
                     log_file_path=log_file_path,
                     db=db,
                 )
@@ -193,6 +201,7 @@ def collect_package_ssh(
                 )
                 _refresh_logs(
                     state_id=state_id,
+                    task_group_activity_id=task_group_activity_id,
                     log_file_path=log_file_path,
                     db=db,
                 )
@@ -202,16 +211,18 @@ def collect_package_ssh(
                 )
                 _refresh_logs(
                     state_id=state_id,
+                    task_group_activity_id=task_group_activity_id,
                     log_file_path=log_file_path,
                     db=db,
                 )
-                stdout_pip_freeze = _customize_and_run_template(
+                _customize_and_run_template(
                     template_name="_4_pip_freeze.sh",
                     **common_args,
                 )
                 logger.debug("installing - END")
                 _refresh_logs(
                     state_id=state_id,
+                    task_group_activity_id=task_group_activity_id,
                     log_file_path=log_file_path,
                     db=db,
                 )
@@ -222,8 +233,10 @@ def collect_package_ssh(
                     logger_name=LOGGER_NAME,
                     db=db,
                 )
+
                 _refresh_logs(
                     state_id=state_id,
+                    task_group_activity_id=task_group_activity_id,
                     log_file_path=log_file_path,
                     db=db,
                 )
@@ -248,6 +261,7 @@ def collect_package_ssh(
 
                 _refresh_logs(
                     state_id=state_id,
+                    task_group_activity_id=task_group_activity_id,
                     log_file_path=log_file_path,
                     db=db,
                 )
@@ -301,19 +315,33 @@ def collect_package_ssh(
                 logger.debug("collecting - END")
                 _refresh_logs(
                     state_id=state_id,
+                    task_group_activity_id=task_group_activity_id,
                     log_file_path=log_file_path,
                     db=db,
                 )
 
                 # Finalize (write metadata to DB)
                 logger.debug("finalising - START")
+                _set_collection_state_data_status(
+                    state_id=state_id,
+                    new_status=CollectionStatusV2.OK,
+                    logger_name=LOGGER_NAME,
+                    db=db,
+                )
+                _set_task_group_activity_status(
+                    task_group_activity_id=task_group_activity_id,
+                    new_status=TaskGroupActivityStatusV2.OK,
+                    logger_name=LOGGER_NAME,
+                    db=db,
+                )
 
-                collection_state = db.get(CollectionStateV2, state_id)
-                collection_state.data["log"] = log_file_path.open("r").read()
-                collection_state.data["freeze"] = stdout_pip_freeze
-                collection_state.data["status"] = CollectionStatusV2.OK
-                flag_modified(collection_state, "data")
-                db.commit()
+                _refresh_logs(
+                    state_id=state_id,
+                    task_group_activity_id=task_group_activity_id,
+                    log_file_path=log_file_path,
+                    db=db,
+                )
+
                 logger.debug("finalising - END")
                 logger.debug("END")
 
@@ -343,6 +371,7 @@ def collect_package_ssh(
                     )
                 _handle_failure(
                     state_id=state_id,
+                    task_group_activity_id=task_group_activity_id,
                     log_file_path=log_file_path,
                     logger_name=LOGGER_NAME,
                     exception=collection_e,