+++ fractal_server/tasks/v2/utils_background.py
@@ -6,9 +6,11 @@ from sqlalchemy.orm.attributes import flag_modified
 from sqlmodel import select
 
 from fractal_server.app.models.v2 import CollectionStateV2
+from fractal_server.app.models.v2 import TaskGroupActivityV2
 from fractal_server.app.models.v2 import TaskGroupV2
 from fractal_server.app.schemas.v2 import CollectionStatusV2
 from fractal_server.app.schemas.v2 import TaskCreateV2
+from fractal_server.app.schemas.v2 import TaskGroupActivityStatusV2
 from fractal_server.app.schemas.v2.manifest import ManifestV2
 from fractal_server.logger import get_logger
 from fractal_server.logger import reset_logger_handlers
@@ -29,18 +31,21 @@ def _set_collection_state_data_status(
     db.commit()
 
 
-def _set_collection_state_data_log(
+def _set_task_group_activity_status(
     *,
-    state_id: int,
-    new_log: str,
+    task_group_activity_id: int,
+    new_status: TaskGroupActivityStatusV2,
     logger_name: str,
     db: DBSyncSession,
 ):
     logger = get_logger(logger_name)
-    logger.debug(f"{state_id=} - set state.data['log']")
-    collection_state = db.get(CollectionStateV2, state_id)
-    collection_state.data["log"] = new_log
-    flag_modified(collection_state, "data")
+    logger.debug(
+        f"{task_group_activity_id=} "
+        "- set task_group_activity.status to {new_status}"
+    )
+    task_group_activity = db.get(TaskGroupActivityV2, task_group_activity_id)
+    task_group_activity.status = TaskGroupActivityStatusV2(new_status)
+    db.add(task_group_activity)
     db.commit()
 
 
@@ -61,6 +66,7 @@ def _set_collection_state_data_info(
 
 def _handle_failure(
     state_id: int,
+    task_group_activity_id: int,
     logger_name: str,
     exception: Exception,
     db: DBSyncSession,
@@ -77,14 +83,20 @@ def _handle_failure(
         db=db,
     )
 
-    new_log = log_file_path.open("r").read()
-
-    _set_collection_state_data_log(
-        state_id=state_id,
-        new_log=new_log,
+    _set_task_group_activity_status(
+        task_group_activity_id=task_group_activity_id,
+        new_status=TaskGroupActivityStatusV2.FAILED,
         logger_name=logger_name,
         db=db,
     )
+
+    _refresh_logs(
+        state_id=state_id,
+        task_group_activity_id=task_group_activity_id,
+        log_file_path=log_file_path,
+        db=db,
+    )
+
     # For backwards-compatibility, we also set state.data["info"]
     _set_collection_state_data_info(
         state_id=state_id,
@@ -109,6 +121,22 @@ def _handle_failure(
         collection_state.taskgroupv2_id = None
         db.add(collection_state)
     logger.info("End of CollectionStateV2 cascade operations.")
+
+    logger.info("Start of TaskGroupActivityV2 cascade operations.")
+    stm = select(TaskGroupActivityV2).where(
+        TaskGroupActivityV2.taskgroupv2_id == task_group_id
+    )
+    res = db.execute(stm)
+    task_group_activity_list = res.scalars().all()
+    for task_group_activity in task_group_activity_list:
+        logger.info(
+            f"Setting TaskGroupActivityV2[{task_group_activity.id}]"
+            ".taskgroupv2_id to None."
+        )
+        task_group_activity.taskgroupv2_id = None
+        db.add(task_group_activity)
+    logger.info("End of TaskGroupActivityV2 cascade operations.")
+
     task_group = db.get(TaskGroupV2, task_group_id)
     db.delete(task_group)
     db.commit()
@@ -197,6 +225,7 @@ def check_task_files_exist(task_list: list[TaskCreateV2]) -> None:
 def _refresh_logs(
     *,
     state_id: int,
+    task_group_activity_id: int,
     log_file_path: Path,
     db: DBSyncSession,
 ) -> None:
@@ -205,5 +234,8 @@ def _refresh_logs(
     """
     collection_state = db.get(CollectionStateV2, state_id)
     collection_state.data["log"] = log_file_path.open("r").read()
+    task_group_activity = db.get(TaskGroupActivityV2, task_group_activity_id)
+    task_group_activity.log = log_file_path.open("r").read()
     flag_modified(collection_state, "data")
+    db.add(task_group_activity)
     db.commit()