+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 **Note**: Numbers like (\#1234) point to closed Pull Requests on the fractal-server repository.
 
+# Unreleased
+
+* API
+    * Add `GET /api/v2/task-group/activity/`. (\#2005)
+    * Add `GET /api/v2/task-group/activity/{task_group_activity_id}/`. (\#2005)
+    * Add `GET /admin/v2/task-group/activity/`. (\#2005)
+* Database
+    * Add `TaskGroupActivityV2` table. (\#2005)
+* Schemas
+    * Add `TaskGroupActivityV2Read`. (\#2005)
+
 # 2.8.1
 
 * API: